src/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
ffmpeg==1.4
ffprobe==0.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
googletrans==4.0.0rc1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.13.7
httpx==1.0.0b0
huggingface-hub==0.24.6
huggingsound==0.1.6
hyperframe==5.2.0
idna==2.10
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiwer==2.6.0
joblib==1.4.2
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
librosa==0.9.2
llvmlite==0.43.0
marisa-trie==1.2.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==0.27.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
pyarrow==17.0.0
pycparser==2.22
pydub==0.25.1
python-dateutil==2.9.0.post0
pytubefix==6.13.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==2.13.7
regex==2024.7.24
requests==2.32.3
resampy==0.4.3
rfc3986==1.5.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.5.0
yarl==1.9.4
scikit-image~=0.24.0
opencv-python~=4.10.0.84
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
